requirements.txt
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.14.2
billiard==4.2.2
celery==5.5.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
Django==5.2.7
django-bootstrap-v5==1.0.11
django-filter==25.2
django-phonenumber-field==8.1.0
django-phonenumbers==1.0.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.1
ruff==0.13.2
service-identity==24.2.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tomli==2.3.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
zope.interface==8.0.1
